+++ core/renderer/ui_component/list/batch_list_adapter.h
@@ -93,7 +93,7 @@ class BatchListAdapter : public ListAdapter {
   void OnItemHolderUpdateFrom(ItemHolder* item_holder) override {}
 
   // Handle diff update to
-  void OnItemHolderUpdateTo(ItemHolder* item_holder) override;
+  void OnItemHolderUpdateTo(ItemHolder* item_holder, bool fiber_flush) override;
 
   // Handle diff moved from
   void OnItemHolderMovedFrom(ItemHolder* item_holder) override {}
@@ -103,6 +103,8 @@ class BatchListAdapter : public ListAdapter {
 
   // Handle diff remove and insert again.
   void OnItemHolderReInsert(ItemHolder* item_holder) override;
+
+  void OnEnqueueElement(ItemHolder* item_holder) override;
 
 #if ENABLE_TRACE_PERFETTO
   void UpdateTraceDebugInfo(TraceEvent* event,